+++ server/src/main/java/org/opensearch/search/internal/ContextIndexSearcher.java
@@ -577,7 +577,24 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
      */
     @Override
     protected LeafSlice[] slices(List<LeafReaderContext> leaves) {
-        return slicesInternal(leaves, searchContext.getTargetMaxSliceCount());
+        if (leaves == null || leaves.isEmpty()) {
+            return new LeafSlice[0];
+        }
+        int targetMaxSlice = searchContext.getTargetMaxSliceCount();
+        if (targetMaxSlice == 0) {
+            LeafSlice[] leafSlices = super.slices(leaves);
+            logger.debug("Slice count using lucene default [{}]", leafSlices.length);
+            return leafSlices;
+        }
+        LeafSlice[] leafSlices = MaxTargetSliceSupplier.getSlices(
+            leaves,
+            targetMaxSlice,
+            searchContext.shouldUseIntraSegmentSearch(),
+            searchContext.getPartitionStrategy(),
+            searchContext.getPartitionMinSegmentSize()
+        );
+        logger.debug("Slice count using max target slice supplier [{}]", leafSlices.length);
+        return leafSlices;
     }
 
     public DirectoryReader getDirectoryReader() {
@@ -644,20 +661,5 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
             }
         }
         return true;
-    }
-
-    // package-private for testing
-    LeafSlice[] slicesInternal(List<LeafReaderContext> leaves, int targetMaxSlice) {
-        LeafSlice[] leafSlices;
-        if (targetMaxSlice == 0) {
-            // use the default lucene slice calculation
-            leafSlices = super.slices(leaves);
-            logger.debug("Slice count using lucene default [{}]", leafSlices.length);
-        } else {
-            // use the custom slice calculation based on targetMaxSlice
-            leafSlices = MaxTargetSliceSupplier.getSlices(leaves, targetMaxSlice);
-            logger.debug("Slice count using max target slice supplier [{}]", leafSlices.length);
-        }
-        return leafSlices;
     }
 }